# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Medical Coding AI",
  "description": "AI-assisted medical coding: clinical documents -> evidence-backed codes -> validation -> human review.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/documents": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Upload a clinical document (PDF, DOCX, TXT)",
    "operationId": "upload_document_api_v1_documents_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProcessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "List Documents",
    "operationId": "list_documents_api_v1_documents_get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "review_route",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Review Route"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DocumentList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/text": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Submit clinical text directly (JSON)",
    "operationId": "submit_text_api_v1_documents_text_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TextDocumentIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProcessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ]
   }
  },
  "/api/v1/documents/{doc_id}": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Get Document",
    "operationId": "get_document_api_v1_documents__doc_id__get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DocumentDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/extract": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Clinical information extraction",
    "operationId": "extract_api_v1_documents__doc_id__extract_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     },
     {
      "name": "async",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Async"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProcessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/code": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "RAG + LLM coding + validation + confidence",
    "operationId": "code_api_v1_documents__doc_id__code_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     },
     {
      "name": "async",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Async"
      }
     },
     {
      "name": "callback_url",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Callback Url"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProcessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/process": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Extraction + coding in one call",
    "operationId": "process_api_v1_documents__doc_id__process_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     },
     {
      "name": "async",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Async"
      }
     },
     {
      "name": "callback_url",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Callback Url"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProcessOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/suggestions": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Suggestions",
    "operationId": "suggestions_api_v1_documents__doc_id__suggestions_get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     },
     {
      "name": "include_superseded",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Superseded"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SuggestionOut"
         },
         "title": "Response Suggestions Api V1 Documents  Doc Id  Suggestions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Coder adds a code the AI missed",
    "operationId": "add_suggestion_api_v1_documents__doc_id__suggestions_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AddCodeIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuggestionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/finalize": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Finalize",
    "operationId": "finalize_api_v1_documents__doc_id__finalize_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FinalCodeOut"
         },
         "title": "Response Finalize Api V1 Documents  Doc Id  Finalize Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/reopen": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Reopen",
    "operationId": "reopen_api_v1_documents__doc_id__reopen_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReopenIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DocumentSummary"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/audit": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Document Audit",
    "operationId": "document_audit_api_v1_documents__doc_id__audit_get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AuditOut"
         },
         "title": "Response Document Audit Api V1 Documents  Doc Id  Audit Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/reviews": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Document Reviews",
    "operationId": "document_reviews_api_v1_documents__doc_id__reviews_get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ReviewOut"
         },
         "title": "Response Document Reviews Api V1 Documents  Doc Id  Reviews Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/runs": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Document Runs",
    "operationId": "document_runs_api_v1_documents__doc_id__runs_get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ModelRunOut"
         },
         "title": "Response Document Runs Api V1 Documents  Doc Id  Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{doc_id}/export": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Export final codes (JSON or CSV)",
    "operationId": "export_api_v1_documents__doc_id__export_get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Doc Id"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "json",
        "csv"
       ],
       "type": "string",
       "default": "json",
       "title": "Format"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_upload_document_api_v1_documents_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     },
     "external_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "External Id"
     },
     "encounter_type": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "outpatient",
         "inpatient"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Encounter Type"
     },
     "patient_sex": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "M",
         "F",
         "U"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Patient Sex"
     },
     "patient_age": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 125.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Patient Age"
     },
     "process": {
      "type": "boolean",
      "title": "Process",
      "default": false
     },
     "callback_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Callback Url"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_document_api_v1_documents_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ProcessOut": {
    "properties": {
     "document_id": {
      "type": "string",
      "format": "uuid",
      "title": "Document Id"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "suggestions": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Suggestions"
     },
     "review_route": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Review Route"
     },
     "by_route": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "integer"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "By Route"
     }
    },
    "type": "object",
    "required": [
     "document_id",
     "status"
    ],
    "title": "ProcessOut"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   },
   "APIKeyHeader": {
    "type": "apiKey",
    "in": "header",
    "name": "X-API-Key"
   }
  }
 }
}